requirements.txt
django>=5.0.0
django_extensions>=3.2.0
python-dateutil>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
